finish rustlings

+++ exercises/23_conversions/using_as.rs
@@ -2,10 +2,12 @@
 // Note that the `as` operator is not only used when type casting. It also helps
 // with renaming imports.
 
+use core::f64;
+
 fn average(values: &[f64]) -> f64 {
     let total = values.iter().sum::<f64>();
     // TODO: Make a conversion before dividing.
-    total / values.len()
+    total / values.len() as f64
 }
 
 fn main() {
